+++ torchdynamo/bytecode_transformation.py
@@ -1,58 +1,11 @@
-import dataclasses
 import dis
 import itertools
 import sys
 import types
-from typing import Any
 from typing import List
-from typing import Optional
 
 from .bytecode_analysis import stacksize_analysis
-
-
-@dataclasses.dataclass
-class Instruction:
-    """A mutable version of dis.Instruction"""
-
-    opcode: int
-    opname: str
-    arg: int
-    argval: Any
-    offset: Optional[int] = None
-    starts_line: Optional[int] = None
-    is_jump_target: bool = False
-    # extra fields to make modification easier:
-    target: Optional["Instruction"] = None
-
-    def __hash__(self):
-        return id(self)
-
-    def __eq__(self, other):
-        return id(self) == id(other)
-    
-    @staticmethod
-    def from_dis(i: dis.Instruction):
-        return Instruction(
-            i.opcode,
-            i.opname,
-            i.arg,
-            i.argval,
-            i.offset,
-            i.starts_line,
-            i.is_jump_target,
-        )
-
-
-class _NotProvided:
-    pass
-
-
-def create_instruction(name, arg=None, argval=_NotProvided, target=None):
-    if argval is _NotProvided:
-        argval = arg
-    return Instruction(
-        opcode=dis.opmap[name], opname=name, arg=arg, argval=argval, target=target
-    )
+from .ir import Instruction, create_instruction
 
 
 def lnotab_writer(lineno, byteno=0):
