+++ torchdynamo/ir/__init__.py
@@ -1,0 +1,2 @@
+from .instruction import Instruction
+from .instruction import create_instruction
